restructre rb ldap and dcu ldap in to seperate sctructs
